+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "moraliser",
-    "version": "1.0.0",
+    "version": "0.2.0",
     "description": "A simple chat software with adult and slang filter",
     "repository": {
         "type": "git",
